pyproject.toml
[project]
name = "clip-assistent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.95.0",
    "python-dotenv>=1.1.1",
]
